+++ src/main/java/com/tavemakers/surf/domain/activity/dto/request/ActivityRecordReqDTO.java
@@ -1,4 +1,4 @@
-package com.tavemakers.surf.domain.activity.dto.reqeust;
+package com.tavemakers.surf.domain.activity.dto.request;
 
 import jakarta.validation.constraints.NotNull;
